Update URIs
* Moved the category summary routes
