+++ app/Http/Controllers/SummaryController.php
@@ -11,7 +11,6 @@ use App\Models\Transformers\ItemYearSummary as ItemYearSummaryTransformer;
 use App\Utilities\Response as UtilityResponse;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Config;
 
 /**
  * Resource summary
@@ -64,283 +63,6 @@ class SummaryController extends Controller
         $routes = [
             'GET' => [
                 'description' => trans('route-descriptions.summary_GET_tco'),
-                'authenticated' => false,
-                'parameters' => []
-            ]
-        ];
-
-        $this->optionsResponse($routes);
-    }
-
-    /**
-     * Return the categories summary for a resource
-     *
-     * @param Request $request
-     * @param string $resource_type_id
-     * @param string $resource_id
-     *
-     * @return JsonResponse
-     */
-    public function categories(
-        Request $request,
-        string $resource_type_id,
-        string $resource_id
-    ): JsonResponse
-    {
-        Validate::resourceRoute($resource_type_id, $resource_id);
-
-        $summary = (new Item())->categoriesSummary(
-            $resource_type_id,
-            $resource_id
-        );
-
-        $headers = [
-            'X-Total-Count' => count($summary)
-        ];
-
-        return response()->json(
-            $summary->map(
-                function ($category_summary) {
-                    return (new ItemCategorySummaryTransformer($category_summary))->toArray();
-                }
-            ),
-            200,
-            $headers
-        );
-    }
-
-    /**
-     * Generate the OPTIONS request for the categories summary
-     *
-     * @param Request $request
-     * @param string $resource_type_id
-     * @param string $resource_id
-     */
-    public function optionsCategories(Request $request, string $resource_type_id, string $resource_id)
-    {
-        Validate::resourceRoute($resource_type_id, $resource_id);
-
-        $routes = [
-            'GET' => [
-                'description' => trans('route-descriptions.summary_GET_categories'),
-                'authenticated' => false,
-                'parameters' => []
-            ]
-        ];
-
-        $this->optionsResponse($routes);
-    }
-
-    /**
-     * Return the resource category summary for a specific category
-     *
-     * @param Request $request
-     * @param string $resource_type_id
-     * @param string $resource_id
-     * @param string $category_id
-     *
-     * @return JsonResponse
-     */
-    public function category(
-        Request $request,
-        string $resource_type_id,
-        string $resource_id,
-        string $category_id
-    ): JsonResponse
-    {
-        Validate::resourceRoute($resource_type_id, $resource_id);
-
-        Validate::categoryRoute($category_id);
-
-        $category_summary = (new Item())->categorySummary(
-            $resource_type_id,
-            $resource_id,
-            $category_id
-        );
-
-        if (count($category_summary) !== 1) {
-            UtilityResponse::notFound();
-        }
-
-        $headers = [
-            'X-Total-Count' => 1
-        ];
-
-        return response()->json(
-            (new ItemCategorySummaryTransformer($category_summary[0]))->toArray(),
-            200,
-            $headers
-        );
-    }
-
-    /**
-     * Generate the OPTIONS request for a category summary
-     *
-     * @param Request $request
-     * @param string $resource_type_id
-     * @param string $resource_id
-     * @param string $category_id
-     */
-    public function optionsCategory(
-        Request $request,
-        string $resource_type_id,
-        string $resource_id,
-        string $category_id
-    ) {
-        Validate::resourceRoute($resource_type_id, $resource_id);
-
-        Validate::categoryRoute($category_id);
-
-        $routes = [
-            'GET' => [
-                'description' => trans('route-descriptions.summary_GET_category'),
-                'authenticated' => false,
-                'parameters' => []
-            ]
-        ];
-
-        $this->optionsResponse($routes);
-    }
-
-    /**
-     * Return the sub categories summary for a resource and category
-     *
-     * @param Request $request
-     * @param string $resource_type_id
-     * @param string $resource_id
-     * @param string $category_id
-     *
-     * @return JsonResponse
-     */
-    public function subCategories(
-        Request $request,
-        string $resource_type_id,
-        string $resource_id,
-        string $category_id
-    ): JsonResponse
-    {
-        Validate::resourceRoute($resource_type_id, $resource_id);
-
-        Validate::categoryRoute($category_id);
-
-        $summary = (new Item())->subCategoriesSummary(
-            $resource_type_id,
-            $resource_id,
-            $category_id
-        );
-
-        $headers = [
-            'X-Total-Count' => count($summary)
-        ];
-
-        return response()->json(
-            $summary->map(
-                function ($category_summary) {
-                    return (new ItemSubCategorySummaryTransformer($category_summary))->toArray();
-                }
-            ),
-            200,
-            $headers
-        );
-    }
-
-    /**
-     * Generate the OPTIONS request for the sub categories summary
-     *
-     * @param Request $request
-     * @param string $resource_type_id
-     * @param string $resource_id
-     * @param string $category_id
-     */
-    public function optionsSubCategories(
-        Request $request,
-        string $resource_type_id,
-        string $resource_id,
-        string $category_id
-    ) {
-        Validate::resourceRoute($resource_type_id, $resource_id);
-
-        Validate::categoryRoute($category_id);
-
-        $routes = [
-            'GET' => [
-                'description' => trans('route-descriptions.summary_GET_sub_categories'),
-                'authenticated' => false,
-                'parameters' => []
-            ]
-        ];
-
-        $this->optionsResponse($routes);
-    }
-
-    /**
-     * Return the resource sub category summary for a specific sub category
-     *
-     * @param Request $request
-     * @param string $resource_type_id
-     * @param string $resource_id
-     * @param string $category_id
-     * @param string $sub_category_id
-     *
-     * @return JsonResponse
-     */
-    public function subCategory(
-        Request $request,
-        string $resource_type_id,
-        string $resource_id,
-        string $category_id,
-        string $sub_category_id
-    ): JsonResponse
-    {
-        Validate::resourceRoute($resource_type_id, $resource_id);
-
-        Validate::subCategoryRoute($category_id, $sub_category_id);
-
-        $sub_category_summary = (new Item())->subCategorySummary(
-            $resource_type_id,
-            $resource_id,
-            $category_id,
-            $sub_category_id
-        );
-
-        if (count($sub_category_summary) !== 1) {
-            UtilityResponse::notFound();
-        }
-
-        $headers = [
-            'X-Total-Count' => 1
-        ];
-
-        return response()->json(
-            (new ItemSubCategorySummaryTransformer($sub_category_summary[0]))->toArray(),
-            200,
-            $headers
-        );
-    }
-
-    /**
-     * Generate the OPTIONS request for the sub category summary
-     *
-     * @param Request $request
-     * @param string $resource_type_id
-     * @param string $resource_id
-     * @param string $category_id
-     * @param string $sub_category_id
-     */
-    public function optionsSubCategory(
-        Request $request,
-        string $resource_type_id,
-        string $resource_id,
-        string $category_id,
-        string $sub_category_id
-    ) {
-        Validate::resourceRoute($resource_type_id, $resource_id);
-
-        Validate::subCategoryRoute($category_id, $sub_category_id);
-
-        $routes = [
-            'GET' => [
-                'description' => trans('route-descriptions.summary_GET_sub_category'),
                 'authenticated' => false,
                 'parameters' => []
             ]
